+++ benches/state_machine.rs
@@ -12,7 +12,7 @@ use failsafe::{backoff, clock, failure_policy, NoopInstrument, StateMachine};
 fn consecutive_failures_policy(b: &mut test::Bencher) {
     let backoff = backoff::constant(Duration::from_secs(5));
     let policy = failure_policy::consecutive_failures(3, backoff);
-    let mut state_machine = StateMachine::new(policy, NoopInstrument);
+    let state_machine = StateMachine::new(policy, NoopInstrument);
 
     b.iter(move || {
         test::black_box(state_machine.is_call_permitted());
@@ -26,7 +26,7 @@ fn success_rate_over_time_window_policy(b: &mut test::Bencher) {
     let backoff = backoff::constant(Duration::from_secs(5));
     let policy =
         failure_policy::success_rate_over_time_window(0.5, 0, Duration::from_secs(10), backoff);
-    let mut state_machine = StateMachine::new(policy, NoopInstrument);
+    let state_machine = StateMachine::new(policy, NoopInstrument);
 
     clock::freeze(|time| {
         b.iter(move || {