+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 Breaking changes:
 * `success_rate` policy now accepts `min_request_threshold`.
+* the `CircuitBreaker` turned into a trait which implements `call` and `call_with` methods.
+* the trait `Callable` was removed
 
 Improvements:
-* add `CircuitBreaker::default()`
 * remove `tokio-timer` dependency.